Notification security/validation and mr suggestions

See merge request cse2115/2022-2023/SEM33a!18

+++ notification-microservice/src/main/java/com/example/notificationmicroservice/controllers/NotificationController.java
@@ -1,7 +1,9 @@
 package com.example.notificationmicroservice.controllers;
 
+import com.example.notificationmicroservice.authentication.AuthManager;
 import com.example.notificationmicroservice.domain.Notification;
 import com.example.notificationmicroservice.services.NotificationDatabaseService;
+import com.example.notificationmicroservice.strategy.NotificationStrategy;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,6 +18,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -24,11 +27,20 @@ import java.util.Map;
 public class NotificationController {
 
     private final transient NotificationDatabaseService notificationDatabaseService;
-    //private final transient AuthService
+    private final transient AuthManager authManager;
+    private transient NotificationStrategy strategy;
 
+    /** Notification controller for injection.
+     *
+     * @param notificationDatabaseService service that communicates with the notification database
+     * @param authManager authentication manager
+     */
     @Autowired
-    public NotificationController(NotificationDatabaseService notificationDatabaseService) {
+    public NotificationController(NotificationDatabaseService notificationDatabaseService,
+                                  AuthManager authManager, NotificationStrategy strategy) {
         this.notificationDatabaseService = notificationDatabaseService;
+        this.authManager = authManager;
+        this.strategy = strategy;
     }
 
     @GetMapping("/hello")
@@ -42,17 +54,32 @@ public class NotificationController {
      * @param targetId the requesting user that wants to get notifications
      * @return notifications addressed to the user
      */
-
     @GetMapping("/getNotifications/{targetId}")
-    public ResponseEntity<List<Notification>> getNotificationsByTarget(@PathVariable String targetId) {
+    public ResponseEntity<List<String>> getNotificationsByTarget(@PathVariable String targetId) {
+        if (!authManager.getNetId().equals(targetId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         List<Notification> notifications = notificationDatabaseService.findNotificationsByTargetId(targetId);
+        List<String> notificationMessages = new ArrayList<>();
+        for (Notification notification : notifications) {
+            notificationMessages.add(notification.buildMessage());
+        }
         notificationDatabaseService.removeNotificationsByTargetId(targetId);
-        return ResponseEntity.ok(notifications);
+        return ResponseEntity.ok(notificationMessages);
     }
 
+    /** Handles the incoming request to notify user with current selected strategy.
+     *
+     * @param notification notification that is to be handled by notification strategy
+     * @return if it succeeded to be handled by notification strategy
+     */
     @PostMapping("/notifyUser")
-    public ResponseEntity<Notification> saveNotification(@RequestBody Notification notification) {
-        return ResponseEntity.ok(notificationDatabaseService.save(notification));
+    public ResponseEntity<String> notifyUser(@RequestBody @Valid Notification notification) {
+        if (!strategy.handleNotification(notification)) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(strategy.getFailureMessage());
+        }
+
+        return ResponseEntity.ok("Successfully notified");
     }
 
     /**
@@ -74,4 +101,7 @@ public class NotificationController {
         return errors;
     }
 
+    public void setNotificationStrategy(NotificationStrategy notificationStrategy) {
+        this.strategy = notificationStrategy;
+    }
 }
